Fix container no not retaining in get by id

+++ src/main/java/com/asg/shipping/importmanifestbl/controller/ImportManifestController.java
@@ -231,7 +231,7 @@ public class ImportManifestController {
     })
     @GetMapping("/load-email-fax")
     public ResponseEntity<?> loadEmailFax(
-            @RequestParam BigDecimal addressMasterPoid,
+            @RequestParam Long addressMasterPoid,
             @RequestParam String addressType
     ) {
             LoadEmailFaxResponseDto response = importManifestService.loadEmailFax(addressMasterPoid, addressType);
